feat: add JWT to requests performed by RestClient instances used in projects and tasks services.
Create JwtInterceptor a custom ClientHttpRequestInterceptor which intercepts request performed by RestClient to add the JWT.
Create JwtAuthenticationToken which extends UsernamePasswordAuthenticationToken to store the JWT.
Modify the JwtAuthenticationFilter to set up the JwtAuthenticationToken instead of UsernamePasswordAuthenticationToken.
Create RestClientConfiguration to properly configure the RestClient builder with the JWT interceptor.
Create an appropriate IT to ensure request done by RestClient are intercepted and modified with the JWT.